refactor: token & strategy model names to kebab case.
