+++ src/services/account-confirmation.service.ts
@@ -1,6 +1,6 @@
 import { autoInjectable, singleton } from "tsyringe";
 import { getRepository, Repository } from "typeorm";
-import AccountConfirmation from "../models/accountConfirmationToken.model";
+import AccountConfirmation from "../models/account-confirmation-token.model";
 import jwt from "jsonwebtoken";
 import { EnvService } from "./env.service";
